separate OCL projects
