+++ ocl_device_utils/ocl_device_utils.h
@@ -20,6 +20,7 @@ public:
 	static bool QueryDevices();
 	static void PrintDevices();
 	static int GetCountForPlatform(int platformID);
+	static void print_opencl_devices();
 
 private:
 	static std::vector<cl::Device> getDevices(std::vector<cl::Platform> const& _platforms, unsigned _platformId);
