+++ ocl_device_utils/opencl.cpp
