+++ ocl_silentarmy/ocl_silentarmy.cpp
