+++ ocl_silentarmy/ocl_silentarmy.hpp
