+++ ocl_silentarmy/ocl_silentarmy.vcxproj.filters
@@ -1,0 +1,28 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <ClCompile Include="ocl_silentarmy.cpp" />
+    <ClCompile Include="sa_blake.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="ocl_silentarmy.hpp" />
+    <ClInclude Include="param.h" />
+    <ClInclude Include="sa_blake.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <Filter Include="zcash">
+      <UniqueIdentifier>{34381c66-ca5c-4daa-aa30-58dcf33e2d66}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="zcash\gpu">
+      <UniqueIdentifier>{c7687099-e206-4d36-8836-f7032bffc7da}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="zcash\gpu\input.cl">
+      <Filter>zcash\gpu</Filter>
+    </None>
+    <None Include="zcash\gpu\kernel.cl">
+      <Filter>zcash\gpu</Filter>
+    </None>
+  </ItemGroup>
+</Project>
