+++ ocl_silentarmy/sa_blake.cpp
