+++ ocl_silentarmy/sa_blake.h
