+++ ocl_silentarmy/zcash/gpu/input.cl
