+++ ocl_silentarmy/zcash/gpu/kernel.cl
